+++ song-server/src/main/java/bio/overture/song/server/security/SystemSecurity.java
@@ -23,7 +23,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 
-import static bio.overture.song.server.security.Scopes.extractGrantedScopes;
+import static bio.overture.song.server.utils.Scopes.extractGrantedScopes;
 
 @Slf4j
 @Component
